added lab4 hints - go checko it out
