+++ playground/micahj/customer.csv
@@ -1,0 +1,3 @@
+name,category
+billbo,awesome
+jimbob, super awesome
